use embedded k0s version for linux installs
