+++ api/internal/managers/app/release/util.go
@@ -5,10 +5,53 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/replicatedhq/embedded-cluster/api/internal/clients"
+	"github.com/replicatedhq/embedded-cluster/pkg/helm"
 	kotsv1beta2 "github.com/replicatedhq/kotskinds/apis/kots/v1beta2"
 	troubleshootv1beta2 "github.com/replicatedhq/troubleshoot/pkg/apis/troubleshoot/v1beta2"
 	"helm.sh/helm/v3/pkg/chart/loader"
 )
+
+func (m *appReleaseManager) setupHelmClient() error {
+	if m.hcli != nil {
+		return nil
+	}
+
+	k8sVersion := m.k8sVersion
+	if k8sVersion == "" {
+		var err error
+		k8sVersion, err = m.getK8sVersion()
+		if err != nil {
+			return fmt.Errorf("get k8s version: %w", err)
+		}
+	}
+
+	hcli, err := helm.NewClient(helm.HelmOptions{
+		// hcli.Render doesn't need a kubeconfig as it is client only
+		K8sVersion: k8sVersion,
+	})
+	if err != nil {
+		return fmt.Errorf("create helm client: %w", err)
+	}
+	m.hcli = hcli
+	return nil
+}
+
+// getK8sVersion creates a kubernetes client and returns the kubernetes version
+func (m *appReleaseManager) getK8sVersion() (string, error) {
+	kcli, err := clients.NewDiscoveryClient(clients.KubeClientOptions{
+		RESTClientGetter: m.restClientGetter,
+		KubeConfigPath:   m.kubeConfigPath,
+	})
+	if err != nil {
+		return "", fmt.Errorf("create discovery client: %w", err)
+	}
+	version, err := kcli.ServerVersion()
+	if err != nil {
+		return "", fmt.Errorf("get server version: %w", err)
+	}
+	return version.String(), nil
+}
 
 // findChartArchive finds the chart archive that corresponds to the given HelmChart CR
 func findChartArchive(helmChartArchives [][]byte, templatedCR *kotsv1beta2.HelmChart) ([]byte, error) {
